Decoupled manga.js into separate folders, created formatOptions.js for all fetch requests.
