+++ commands.js
@@ -3,17 +3,18 @@ const fetch = require('node-fetch');
 const { SlashCommandBuilder } = require('@discordjs/builders');
 const { MessageButton, MessageEmbed } = require('discord.js');
 const paginationEmbed = require('discordjs-button-pagination');
+const { formatOptions } = require('./options.js');
 
 const {
-  updateMangaList, getMangaId, createList, getFieldsFromMangaIds,
-  getMangaIdsFromList, getListId, getMangaData, getListData
-} = require('./manga.js');
+  updateMangaList, createList, getFieldsFromMangaIds,
+  getMangaIdsFromList, getMangaData, getListData
+} = require('./manga/mgExport.js');
 
 
 const {
   insertFollow, delFollow, getGuildRow, getMangaCount,
   updateChannelId, insertGuildRow, getFollowedMangas, getGuildTable
-} = require('./postgres/postgres.js');
+} = require('./postgres/psExport.js');
 
 
 const followCommand = new SlashCommandBuilder()
@@ -55,18 +56,12 @@ const commands = [followCommand, unfollowCommand, setChannelCommands, listMangaC
 const global_url = `https://discord.com/api/v8/applications/${process.env.APPLICATION_ID}/commands`;
 
 
-function getCurrentCommands() {
-  const options = {
-    method: 'GET',
-    headers: {
-      'Authorization': `Bot ${process.env.DISCORD_TOKEN}`,
-      'Content-Type': 'application/json'
-    }
-  };
+async function getCurrentCommands() {
+  const options = formatOptions('GET', `Bot ${process.env.DISCORD_TOKEN}`);
 
-  return fetch(global_url, options).then(res => res.json()).then(json => {
-    return json.map(command => command.name);
-  }).catch(err => console.log(err));
+  const res = await fetch(global_url, options).catch(err => console.log(err));
+  const json = await res.json();
+  return json.map(command => command.name);
 }
 
 
@@ -74,14 +69,7 @@ async function createCommands() {
   const currents = await getCurrentCommands();
   for (const command of commands) {
     if (currents.includes(command.name)) { continue; }
-    const options = {
-      method: 'POST',
-      body: JSON.stringify(command.toJSON()),
-      headers: {
-        'Authorization': `Bot ${process.env.DISCORD_TOKEN}`,
-        'Content-Type': 'application/json'
-      }
-    };
+    const options = formatOptions('POST', `Bot ${process.env.DISCORD_TOKEN}`, command.toJSON());
     const response = await fetch(global_url, options);
     console.log(await response.json());
   }
@@ -92,6 +80,28 @@ async function createCommands() {
 function checkGuild(guildId) {
   //Returns true if the guild exists in the table, false if not.
   return getGuildRow(guildId).then(res => Boolean(res?.length || 0));
+}
+
+
+function getMangaId(intOptions) {
+  // Get mangaID from the URL, returns empty string if invalid URL.
+  const input = intOptions.getString('url');
+  if (input.slice(0, 27) === 'https://mangadex.org/title/') {
+    return input.slice(27).split('/')[0];
+  }
+  console.log('Invalid URL', input);
+  return '';
+}
+
+
+function getListId(intOptions) {
+  //Grabs listID from listUrl.
+  const input = intOptions.getString('url');
+  if (input.slice(0, 26) !== 'https://mangadex.org/list/') {
+    console.log('Invalid URL', input);
+    return '';
+  }
+  return input.slice(26).split('/')[0];
 }
 
 
@@ -153,8 +163,11 @@ async function handleUnfollowCommand(interaction) {
     const mangaCount = await getMangaCount(mangaId);
     if (mangaCount === '0') {
       const listId = (await getGuildRow(guildId))?.[0]?.list_id;
-      await updateMangaList(mangaId, listId, 'DELETE');
-      console.log(`${mangaTitle || mangaId} was deleted from list: ${listId}`);
+      const res = await updateMangaList(mangaId, listId, 'DELETE');
+      if (res === 'ok') { console.log(`${mangaTitle || mangaId} was deleted from list: ${listId}`) }
+      else { 
+        console.log('Error with list');
+      }
     }
     await interaction.editReply({ content: `No longer following ${mangaTitle || mangaId}` });
   }
@@ -245,8 +258,8 @@ async function handleMigrateCommand(interaction) {
 
   await interaction.editReply({ content: `Migrating ${mangaIds.length} mangas...` });
   for (const mangaId of mangaIds) {
-    await Promise.all([updateMangaList(mangaId, newListId, 'POST'),
-    insertFollow(userId, mangaId, guildId)]);
+    await updateMangaList(mangaId, newListId, 'POST');
+    await insertFollow(userId, mangaId, guildId);
   }
 
   await interaction.editReply({ content: `Migrated ${mangaIds.length} mangas from ${listTitle} to server list.` });
@@ -260,3 +273,4 @@ module.exports = {
   handleListCommand,
   handleMigrateCommand
 };
+
